+++ src/request-scope/request-scope.middleware.ts
@@ -2,7 +2,8 @@ import { createNamespace, Namespace } from 'cls-hooked';
 import { Inject, Injectable, NestMiddleware, Optional } from '@nestjs/common';
 import { Request, Response } from 'express';
 import { _REQUEST_STORAGE_NAMESPACE_KEY } from './request-scope';
-import { rootLogger } from '..';
+import { Configuration, CONFIGURATION, rootLogger } from '..';
+import { root } from 'rxjs/internal-compatibility';
 
 export interface RequestScopeInterceptor {
   /**
@@ -19,19 +20,35 @@ export const REQUEST_SCOPE_INTERCEPTORS = 'REQ_SCOPE_INTERCEPTORS';
 
 @Injectable()
 export class RequestScopeMiddleware implements NestMiddleware {
-  private namespace: Namespace;
-  constructor(@Inject(REQUEST_SCOPE_INTERCEPTORS) private readonly interceptors: RequestScopeInterceptor[]) {
-    this.namespace = createNamespace(_REQUEST_STORAGE_NAMESPACE_KEY);
-    const interceptorNames = interceptors.map((interceptor) => `   - ${interceptor.name}`).join('\n');
-    rootLogger.info(`RequestScopeMiddleware setup up with interceptors: \n${interceptorNames}`);
+  private readonly namespace?: Namespace;
+  private readonly enabled: boolean;
+
+  constructor(
+    @Inject(CONFIGURATION) configurationProvider: Configuration,
+    @Inject(REQUEST_SCOPE_INTERCEPTORS) private readonly interceptors: RequestScopeInterceptor[],
+  ) {
+    // Enabled by default
+    const disabled = configurationProvider.requestScope?.enabled === false;
+    this.enabled = !disabled;
+    if (this.enabled) {
+      this.namespace = createNamespace(_REQUEST_STORAGE_NAMESPACE_KEY);
+      const interceptorNames = interceptors.map((interceptor) => `   - ${interceptor.name}`).join('\n');
+      rootLogger.info(`RequestScopeMiddleware setup up with interceptors: \n${interceptorNames}`);
+    } else {
+      rootLogger.info('RequestScopeMiddleware disabled by config');
+    }
   }
 
   use(req: Request, res: Response, next: () => void) {
-    this.namespace.run(() => {
-      this.interceptors.forEach((interceptor) => {
-        interceptor.intercept(req);
+    if (this.namespace) {
+      this.namespace.run(() => {
+        this.interceptors.forEach((interceptor) => {
+          interceptor.intercept(req);
+        });
+        next();
       });
+    } else {
       next();
-    });
+    }
   }
 }